requirements/base.txt
asgiref==3.2.10
Django==3.1.2
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
django-tinymce
gunicorn